refactor dataset.py classes into seperate files
